Refactor: 서비스 로직에서 커넥션 타입 명시적으로 지정하고, try 블록 내부에서 커넥션을 가져오도록 수정

+++ backend/src/service/trade/delete.ts
@@ -1,14 +1,18 @@
 import db from '@config/database';
+import { PoolConnection } from 'mysql2/promise';
 
 // 교환글 삭제
 export const deleteTrade = async (tradeId: number, voucherId: number) => {
-  const con = await db.getConnection();
+  let con: PoolConnection | undefined;
 
   try {
+    con = await db.getConnection();
     await con.beginTransaction();
 
     // 소유권 상태 변경
     const updateVoucher = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       UPDATE Voucher
       SET
@@ -20,6 +24,8 @@ export const deleteTrade = async (tradeId: number, voucherId: number) => {
 
     // 교환글 삭제
     const deleteTrade = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       DELETE FROM Trade
       WHERE trade_id=${con.escape(tradeId)}`;
@@ -30,9 +36,9 @@ export const deleteTrade = async (tradeId: number, voucherId: number) => {
     await Promise.all([updateVoucher, deleteTrade]);
     con.commit();
   } catch (err) {
-    con.rollback();
+    con?.rollback();
     throw err;
   } finally {
-    con.release();
+    con?.release();
   }
 }
