+++ backend/src/service/trade/update.ts
@@ -1,21 +1,29 @@
-
 import db from '@config/database';
+import { PoolConnection } from 'mysql2/promise';
 import { TradeDetail } from '@type/trade';
 
 // 교환글 수정
-export const updateTrade = async ({ trade, voucherId, amount, wantPhotocardIds }: {
+export const updateTrade = async ({
+  trade, 
+  voucherId, 
+  amount, 
+  wantPhotocardIds
+}: {
   trade: TradeDetail;
   voucherId: number;
   amount: number;
   wantPhotocardIds: number[];
 }) => {
-  const con = await db.getConnection();
+  let con: PoolConnection | undefined;
 
   try {
+    con = await db.getConnection();
     await con.beginTransaction();
 
     // 기존 소유권 상태를 available로 변경
     const updateExistingVoucher = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       UPDATE Voucher
       SET
@@ -27,6 +35,8 @@ export const updateTrade = async ({ trade, voucherId, amount, wantPhotocardIds }
 
     // 새로운 소유권 사용상태 변경
     const updateNewVoucher = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       UPDATE Voucher
       SET
@@ -38,6 +48,8 @@ export const updateTrade = async ({ trade, voucherId, amount, wantPhotocardIds }
 
     // 기존 wantPhotocard 모두 제거
     const deleteExistingWantcard = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+      
       let sql = `
       DELETE FROM TradeWantcard
       WHERE trade_id=${trade.tradeId}`
@@ -47,6 +59,8 @@ export const updateTrade = async ({ trade, voucherId, amount, wantPhotocardIds }
 
     // 교환글 수정
     const updateTrade = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       UPDATE Trade
       SET
@@ -60,6 +74,8 @@ export const updateTrade = async ({ trade, voucherId, amount, wantPhotocardIds }
     // 교환글이 원하는 포토카드 정보 작성
     const insertWantcards = wantPhotocardIds.map(photocardId => (
       new Promise((resolve, reject) => {
+        if (!con) return reject(new Error('undefined db connection'));
+
         let sql = `
         INSERT INTO TradeWantcard(
           trade_id,
@@ -78,31 +94,40 @@ export const updateTrade = async ({ trade, voucherId, amount, wantPhotocardIds }
       updateNewVoucher,
       deleteExistingWantcard,
       updateTrade,
-      ...insertWantcards]);
+      ...insertWantcards
+    ]);
 
     con.commit();
   } catch (err) {
-    con.rollback();
+    con?.rollback();
     throw err;
   } finally {
-    con.release();
+    con?.release();
   }
 }
 
 // 교환 진행
-export const exchangeTrade = async ({ tradeId, voucherId, userId, voucherIds }: {
+export const exchangeTrade = async ({
+  tradeId,
+  voucherId,
+  userId,
+  voucherIds
+}: {
   tradeId: number;
   voucherId: number;
   userId: number;
   voucherIds: number[];
 }) => {
-  const con = await db.getConnection();
+  let con: PoolConnection | undefined;
 
   try {
+    con = await db.getConnection();
     await con.beginTransaction();
 
     // 교환글 상태 변경, 교환일 변경
     const updateTrade = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       UPDATE Trade
       SET
@@ -115,6 +140,8 @@ export const exchangeTrade = async ({ tradeId, voucherId, userId, voucherIds }: 
     
     // 교환글에 등록된 소유권의 주인과 상태 변경
     const updateTradeVoucher = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       UPDATE Voucher
       SET
@@ -127,6 +154,8 @@ export const exchangeTrade = async ({ tradeId, voucherId, userId, voucherIds }: 
 
     // 교환 기록 추가
     const insertVoucherLog = new Promise((resolve, reject) => {
+      if (!con) return reject(new Error('undefined db connection'));
+
       let sql = `
       INSERT INTO VoucherLog(
         voucher_id,
@@ -149,6 +178,8 @@ export const exchangeTrade = async ({ tradeId, voucherId, userId, voucherIds }: 
 
         // 소유권 변경
         const updateVoucher = new Promise((resolve, reject) => {
+          if (!con) return reject(new Error('undefined db connection'));
+
           let sql = `
           UPDATE Voucher
           SET
@@ -161,6 +192,8 @@ export const exchangeTrade = async ({ tradeId, voucherId, userId, voucherIds }: 
 
         // 교환 기록 작성
         const insertLog = new Promise((resolve, reject) => {
+          if (!con) return reject(new Error('undefined db connection'));
+          
           let sql = `
           INSERT INTO VoucherLog(
             voucher_id,
@@ -190,9 +223,9 @@ export const exchangeTrade = async ({ tradeId, voucherId, userId, voucherIds }: 
 
     con.commit();
   } catch (err) {
-    con.rollback();
+    con?.rollback();
     throw err;
   } finally {
-    con.release();
+    con?.release();
   }
 }
